fix wrong date, uuid type and enum in entities
